added personal stats

+++ Trivia-Project/Headers/IDataBase.h
@@ -28,7 +28,7 @@ public:
 	// Statistics
 	virtual int addStatistic(std::string username, std::string time, std::string isCorrectAnswer, std::string score) = 0;
 	virtual QuestionsList* getQuestions(int amount) = 0;
-	virtual int getPlayerAverageAnswerTime(std::string username) = 0;
+	virtual float getPlayerAverageAnswerTime(std::string username) = 0;
 	virtual int getNumOfCorrectAnswers(std::string uaername) = 0;
 	virtual int getNumOfWrongAnswers(std::string username) = 0;
 	virtual int getNumOfPlayerGames(std::string username) = 0;
